project-root/visualization/requirements.txt
flask==3.0.0
pyyaml==6.0.1
requests==2.31.0